+++ simcity.h
@@ -624,7 +624,7 @@ public:
 	void set_citygrowth_yesno( bool ng ) { allow_citygrowth = ng; }
 	bool get_citygrowth() const { return allow_citygrowth; }
 
-	void step(long delta_t);
+	void step(uint32 delta_t);
 
 	void new_month(bool check);
